requirement.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
phonenumbers==8.12.49
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
